Added atomic file writer function to library
The reason that this function DOESN'T auto-guess the filepath
is that that's a major weakness of the OTHER functions in this
library. It means that the user can never create brand new
files/cache files that don't already exist on disk. That's
a huge bug that will need to be FIXED in all OTHER functions.
Possibly by entirely ripping out and throwing away the stupid
auto-guessing behavior in all other functions that use it.