Removed insane code which writes to random files
This was one of the dumbest ideas in the history of programming,
and the original author should be ashamed of himself. The way
it randomly searched all paths for a filename meant that you
never knew where it ended up writing/reading the data from.

Furthermore, the terrible way in which it was written meant that
you could never create any new files in your data/cache folders,
since the code only succeeded if the target file already existed.
So the old code was never able to write anything anyway.

Idiotic.