read(): Rewrote the terribly broken file reader
Renamed the retarded "returnAsObject" which really meant
"returnAsArray". The original author must have misread
the json_decode() function documentation.

Now we call it "getJsonAsArray", and the default is now
TRUE, in order to be consistent with the way the writer
outputs arrays as JSON. The previous default was bugged.
The array, if read back into memory, would always become
an object instead of an array.

Furthermore, the decoding logic was fixed, because it
was a broken mess. Now it efficiently checks if the data
is possibly JSON, and then attempts a decode (which quits
instantly if it isn't JSON after all). And any decoding
errors are silenced, so that the whole magic is transparent
to the end user. So if they've saved JSON, they get an array
or object. Otherwise they get the raw file contents. They
never have to worry about these details manually.

Also fixed all function documentation.