Made the "arg" and "uid" parameters optional
There is no requirement in Alfred that the user provides
these parameters. Therefore, the behavior was changed
as follows:

uid: It could already be set to NULL to omit it completely,
but now we actually include it if we receive an empty string.
That was done for more consistency. Empty strings should
never mean "omit this parameter". That's what NULL is for.

arg: Can now be omitted by passing in NULL. People shouldn't
be forced to provide any argument, since lots of workflow
results don't need to pass any data onwards.