Constructor: Fixed the bundle ID detection logic
The old code was totally broken. If it didn't find "Info.plist",
or couldn't read the bundle ID from it, and the user hadn't provided
any ID override, then it simply happily went onwards with an empty
ID, which meant that the cache/data paths were misgenerated and that
if you tried to write any files to those paths, you would end up
writing to the wrong disk locations.

Furthermore, the old code only used the constructor's own bundleId
parameter if it couldn't extract anything from the plist, which
made the parameter totally useless since there's no point having
an override if you can't override it!

The rewrite now works as follows: If the user has provided a bundleId
override, we select that. Otherwise we look in the workflow's current
working directory and read "info.plist" and attempt to extract the
bundleId from there. If we don't have any bundleId at the end of all
of that, we fall back to using "com.alfredworkflow.unknownbundleid",
which is a generic ID that simply ensures that any data/cache files
you create won't be written to the wrong folders. At least they will
now end up in a nice workflow subfolder!

Note: This code uses the new filepath() function, to be committed next.