Added function for easily generating file paths
Makes it effortless to tell the workflow to save/load data from somewhere.

Just use this function to generate a path which combines your desired
folder (such as "data" or "cache") and the filename you want. The
result will be expanded to the full path to that folder and your file.

The target file doesn't have to exist yet.