Tons of cleanup of old code documentation and style
There were lots of bugs in the old documentation, with it often
claiming things that weren't true about the functions.

Lots of them have now been rewritten to be brought up to date,
cleaned up and clarified.

Also did some general cleanup of coding style. The old code is
still a mess, but it's at least a little bit more readable now.