Remove strict type declarations to support PHP5
PHP5 is terrible, but it's the default included on macOS. So remove
all PHP7 type declarations. It's easy for me to forget the old coding
style since I've personally left the slow, bloated PHP5 behind. ;-)

The PHP7 engine is 6x faster than the PHP5 engine. Too bad it's not
included with macOS yet. At Apple's rate, that'll probably take a decade.